requirements.txt
numpy
opencv-python
pyyaml>=6.0
# picamera2              # Raspberry Pi camera library — enable when running on Pi 5
# pyserial               # Serial comms with robot — enable when hardware is available
